backend/pyproject.toml
[project]
name = "den.bot"
version = "0.1.0"
description = "den.bot website"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-phonenumber-field[phonenumberslite]>=8.1.0",
    "djangorestframework>=3.16.1",
    "python-dotenv>=1.1.0",
    "twilio>=9.8.1",
]
